feat(remote-api): Move che-server implementation classes to its own module, to allow other implementations Part of https://github.com/eclipse/che/issues/18874
