+++ extensions/eclipse-che-theia-remote-api/src/common/index.ts
@@ -1,0 +1,18 @@
+/**********************************************************************
+ * Copyright (c) 2021 Red Hat, Inc.
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ ***********************************************************************/
+
+export * from './certificate-service';
+export * from './factory-service';
+export * from './k8s-service';
+export * from './oauth-service';
+export * from './ssh-key-service';
+export * from './telemetry-service';
+export * from './user-service';
+export * from './workspace-service';
