+++ extensions/eclipse-che-theia-remote-impl-che-server/.gitignore
@@ -1,0 +1,1 @@
+lib
