+++ extensions/eclipse-che-theia-remote-impl-che-server/src/node/che-remote-impl-che-server-backend-module.ts
@@ -7,15 +7,19 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  ***********************************************************************/
-import { CertificateService, cheCertificateServicePath } from '../common/certificate-service';
-import { CheK8SService, cheK8SServicePath } from '../common/k8s-service';
+
+import {
+  CertificateService,
+  cheCertificateServicePath,
+} from '@eclipse-che/theia-remote-api/lib/common/certificate-service';
+import { CheK8SService, cheK8SServicePath } from '@eclipse-che/theia-remote-api/lib/common/k8s-service';
 import { ConnectionHandler, JsonRpcConnectionHandler } from '@theia/core';
-import { FactoryService, cheFactoryServicePath } from '../common/factory-service';
-import { OAuthService, cheOAuthServicePath } from '../common/oauth-service';
-import { SshKeyService, cheSshKeyServicePath } from '../common/ssh-key-service';
-import { TelemetryService, cheTelemetryServicePath } from '../common/telemetry-service';
-import { UserService, cheUserServicePath } from '../common/user-service';
-import { WorkspaceService, cheWorkspaceServicePath } from '../common/workspace-service';
+import { FactoryService, cheFactoryServicePath } from '@eclipse-che/theia-remote-api/lib/common/factory-service';
+import { OAuthService, cheOAuthServicePath } from '@eclipse-che/theia-remote-api/lib/common/oauth-service';
+import { SshKeyService, cheSshKeyServicePath } from '@eclipse-che/theia-remote-api/lib/common/ssh-key-service';
+import { TelemetryService, cheTelemetryServicePath } from '@eclipse-che/theia-remote-api/lib/common/telemetry-service';
+import { UserService, cheUserServicePath } from '@eclipse-che/theia-remote-api/lib/common/user-service';
+import { WorkspaceService, cheWorkspaceServicePath } from '@eclipse-che/theia-remote-api/lib/common/workspace-service';
 
 import { CheK8SServiceImpl } from './che-server-k8s-service-impl';
 import { CheServerCertificateServiceImpl } from './che-server-certificate-service-impl';
